+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 Changes:
 
+- Add internal merge of single storage queries
 - Expose provider stats on API intrerface via `.stats`
 - Adjust flag extraction for latest phragmen types
 - Add Kusama 9220 upgrade block